♻️ Merge the tests to remove duplicated tests and clear the tests.
