+++ test/e2e_test/web_server/test_mount.py
@@ -40,31 +40,3 @@ class TestWebServerMountMcpServer(BaseE2ETestWithRunningServer):
         # If the endpoint is not mounted, we would get a 404 Not Found
         assert status_code != 404, f"Expected status non-(404) for {server_fixture.url_suffix}, got {status_code}"
         assert status_code in (200, 307), f"Expected status (200,307) for {server_fixture.url_suffix}, got {status_code}"
-
-
-    # Test Common Endpoints (regardless of transport)
-
-
-    @pytest.mark.parametrize(
-        "endpoint",
-        [
-            "/health",  # Root health check
-        ],
-    )
-    def test_common_endpoints(self, server_fixture: MCPServerFixtureValue, endpoint: str) -> None:
-        """Test that common endpoints are available regardless of transport type."""
-        base_url = f"http://{server_fixture.host}:{server_fixture.port}"
-        url = f"{base_url}{endpoint}"
-
-        with httpx.Client(timeout=OPERATION_TIMEOUT) as client:
-            response = client.get(url)
-
-        assert response.status_code == 200, f"Expected status 200 for {endpoint}, got {response.status_code}"
-
-        # Verify root endpoint response
-        if endpoint == "/health":
-            json_response = response.json()
-            assert "status" in json_response, "Missing status field in response"
-            assert json_response.get("status") == "ok", "Invalid status in response"
-            assert "server" in json_response, "Missing server field in response"
-            assert json_response.get("server") == "ClickUp MCP Server", "Invalid server in response"
